[mlir][VectorUtils] Add utility for computing scalable value upper bounds
This PR adds `vector::computeScalableUpperBound()` which is built using
the `ValueBoundsConstraintSet`. This function works like a scalable
version of `ValueBoundsConstraintSet::computeConstantBound()`, but
rather than simply returning a constant value it can also return bounds
of the form `vscale * constant`, such bounds are very common when
working with scalable vectors.

The interface is defined as:

```c++
FailureOr<BoundSize> vector::computeScalableUpperBound(
  Value value,
  std::optional<int64_t> dim,
  unsigned vscaleMin = 1,
  unsigned vscaleMax = 16,
  bool vscaleIsPow2 = true);
```

It works by first computing the constraint set assuming `vscale` is
`vscaleMax`, then repeatedly shrinking range of `vscale` until it
reaches `vscaleMin`. If at each step the upper bound shrank
proportionally the bound can be assumed to be
`(bound at vscale == 1) * vscale`. If the upper bound does not follow
this pattern (but can be computed) a conservative fixed-sized bound
(assuming `vscale == 16`) will be returned instead.

This can be computed with `log2(vscaleMax - vscaleMin)` extra
constraints assuming `vscale` is a power of two (which is a fairly
reasonable assumption).