[mlir][vector] Update syntax and representation of insert/extract_strided_slice
This commit updates the representation of both `extract_strided_slice`
and `insert_strided_slice` to primitive arrays of int64_ts, rather than
ArrayAttrs of IntegerAttrs. This prevents a lot of boilerplate
conversions between IntegerAttr and int64_t.

Because previously the offsets, strides, and sizes were in the
attribute dictionary (with no special syntax), simply replacing the
attribute types with `DenseI64ArrayAttr` would be a syntax break.

So since a break is unavoidable this commit also tackles a long-standing
TODO:

```mlir
// TODO: Evolve to a range form syntax similar to:
%1 = vector.extract_strided_slice %0[0:2:1][2:4:1]
  : vector<4x8x16xf32> to vector<2x4x16xf32>
```

This is done by introducing a new `StridedSliceAttr` attribute that can
be used for both operations, with syntax based on the above example.

See the attribute documentation `VectorAttributes.td` for a full
overview.

Automatically upgrade MLIR check tests

All changes in this commit where done automatically with the following
Python script:

https://gist.github.com/MacDue/ca84d3ec19cf83ae71aab2be8f09c3c5

Manually upgrade three tests (with non-standard syntax/checks)