[mlir][vector] Update representation of insert/extract_strided_slice
This commit updates the representation of both extract_strided_slice and
insert_strided_slice to primitive arrays of int64_ts, rather than
ArrayAttrs of IntegerAttrs. This prevents a lot of boilerplate
conversions between IntegerAttr and int64_t.

This is done by adding a new `StridedSliceAttr` which matches the
previous syntax and can be used for both operations.

It may also be possible to explore alternate slice syntax for the
`StridedSliceAttr` in future.