[AArch64][SME] Improve support for agnostic ZA functions
This patch ensures that in agnostic ZA functions, if a callee may throw
an exception, which the current function could catch, we commit a ZA
save before that callee. This ensures we will be able to restore ZA
state if we unwind to the catch block.

This patch also no longer requires ZA to be in the "ACTIVE" state before
agnostic ZA callees (since these functions do not need the contents of
ZA).

Change-Id: I16924e61b1f36695b23a28e4dae7bce003e5711c